add sort min transition

+++ main.py
@@ -1,4 +1,5 @@
 from lib.synopsys import Synopsys
+from lib.sort_min_transition import SortMinTransition
 from string import Template
 import os
 
@@ -25,24 +26,47 @@ def request_SDQL(settings):
     combine = Synopsys.combine('../template/RequestSDQL', settings)
     Synopsys.tmax(combine)
 
-def analysys_power(settings):
-    ## 電力を求める
+def make_dump(settings):
     ## stildpv.v ファイルにdumpファイルを生成するプログラムの追加
+    already = False
     with open(settings["name"] + '_stildpv.v', 'r') as f:
         stil_file = f.readlines()
     for i in stil_file:
         if 'vector_number = 0;' in i:
             index = stil_file.index(i)
-    stil_file.insert(index+1, '$dumpfile("' + settings["name"] + '.vcd");\n')
-    stil_file.insert(index+2, '$dumpvars(0, ' + settings["name"] + ');\n')
-    with open(settings["name"] + '_stildpv.v', 'w') as f:
-        f.writelines(stil_file)
+        if '$dumpfile(' in i:
+            # すでにdumfileが書かれていた場合
+            already = True
+    if already != True:
+        stil_file.insert(index+1, '$dumpfile("' + settings["name"] + '.vcd");\n')
+        stil_file.insert(index+2, '$dumpvars(0, ' + settings["name"] + ');\n')
+        with open(settings["name"] + '_stildpv.v', 'w') as f:
+            f.writelines(stil_file)
 
+def analysys_power(settings):
+    ## 電力を求める
     # vcdファイルを作るためのshを実行
+    make_dump(settings)
     os.system('bash ' + settings["name"] + '_vcs.sh')
 
     combine = Synopsys.combine('../template/AnalysisPower', settings)
     Synopsys.pt_shell(combine)
+
+def analysys_power_f(settings, f):
+    ## ファイル指定したstilで電力を求める
+    os.system('cp ' + settings["stil"] + ' temp.stil')
+    os.system('cp ' + f + ' ' + settings["stil"])
+
+    settings["power"] = f + '_report_power.txt'
+
+    make_dump(settings)
+    os.system('bash ' + settings["name"] + '_vcs.sh')
+
+    combine = Synopsys.combine('../template/AnalysisPower', settings)
+    Synopsys.pt_shell(combine)
+
+    os.system('cp ' + ' temp.stil ' + settings["stil"])
+    os.remove('temp.stil')
 
 def synth_to_SDQL(settings):
     logic_synthesis(settings)
@@ -51,10 +75,11 @@ def synth_to_SDQL(settings):
     request_SDQL(settings)
 
 def clock_judge(target):
-    if target in ["b04", "b05", "b08", "b15", "b17"]:
+    if target in ["b04", "b05", "b08", "b15"]:
         return "CLOCK"
     else:
         return "clock"
+
 
 if __name__ == '__main__':
     target = 'b10'
@@ -73,12 +98,14 @@ if __name__ == '__main__':
                     slk        = target + '.slk',
                     stilcsv    = target + '.stilcsv',
                     vcd        = target + '.vcd',
-                    fault      = target + '_report_faults.txt'
+                    fault      = target + '_report_faults.txt',
+                    power      = target + '_report_power.txt'
                     )
-
 
     synth_to_SDQL(settings)
     analysys_power(settings)
+    SortMinTransition.sort('b04.stil', 'b04_sort.stil')
+    analysys_power_f(settings, 'b10_sorted.stil')
 
     # ローカルにファイル転送
     # os.system('rsync -ruz ../. kazutaka@192.168.7.149:/Users/kazutaka/Git/NNCT_Sudy')
